+++ wemake_python_styleguide/violations/best_practices.py
@@ -3083,6 +3083,22 @@ class MultilineFormattedStringViolation(TokenizeViolation):
     """
     Forbid using multi-line formatted string with single and double quotes.
 
+    Reasoning:
+        Multiline f-strings must use triple quotes for clarity; single f-strings may not span lines.
+
+    Solution:
+        Use  \"\"\" quotes instead of single quotes.
+
+    Example::
+
+        # Correct:
+        x = f\"\"\" { 1
+        ...}\"\"\"
+
+        # Wrong:
+        x = f\" { 1
+        ...}\"
+
     .. versionadded:: 1.2.0
 
     """